feat(nx-docker): improvements in nx-docker package

+++ packages/nx-docker/src/executors/build/main.ts
@@ -2,7 +2,6 @@ import * as core from '@nx-tools/core';
 import { getMetadata } from '@nx-tools/docker-meta';
 import { config } from 'dotenv';
 import * as fs from 'fs';
-import * as os from 'os';
 import * as buildx from './buildx';
 import * as context from './context';
 import * as exec from './exec';
@@ -11,21 +10,17 @@ import { BuildExecutorSchema } from './schema';
 export default async function run(options: BuildExecutorSchema): Promise<{ success: true }> {
   config();
 
-  if (os.platform() !== 'linux' && os.platform() !== 'darwin') {
-    throw new Error(`Only supported on linux and darwin platform`);
-  }
+  core.startGroup(`Docker info`);
+  await exec.exec('docker', ['version']);
+  await exec.exec('docker', ['info']);
+  core.endGroup();
 
   if (!(await buildx.isAvailable())) {
-    throw new Error(`Buildx is required. See https://github.com/docker/setup-buildx-action to set up buildx.`);
+    throw new Error(`Docker buildx is required. See https://github.com/docker/setup-buildx-action to set up buildx.`);
   }
 
   const tmpDir = context.tmpDir();
-
-  core.info(`tmpDir ${tmpDir}`);
-
   const buildxVersion = await buildx.getVersion();
-  core.info(`📣 Buildx version: ${buildxVersion}`);
-
   const defContext = context.defaultContext();
   const inputs: context.Inputs = await context.getInputs(defContext, options);
 
@@ -40,7 +35,7 @@ export default async function run(options: BuildExecutorSchema): Promise<{ succe
     }
   }
 
-  core.info(`🏃 Starting build...`);
+  core.info(`Starting build...`);
   const args: string[] = await context.getArgs(inputs, defContext, buildxVersion);
 
   const dockerCmd = `docker ${args.join(' ')}`;
@@ -55,17 +50,20 @@ export default async function run(options: BuildExecutorSchema): Promise<{ succe
 
   const imageID = await buildx.getImageID();
   if (imageID) {
-    core.info('🛒 Extracting digest...');
+    core.startGroup(`Extracting digest`);
     core.info(`${imageID}`);
+    core.endGroup();
   }
 
-  cleanup();
+  cleanup(tmpDir);
 
   return { success: true };
 }
 
-async function cleanup(): Promise<void> {
-  const tmpDir = context.tmpDir();
-  core.info(`🚿 Removing temp folder ${tmpDir}`);
-  fs.rmdirSync(tmpDir, { recursive: true });
+async function cleanup(tmpDir: string): Promise<void> {
+  if (tmpDir.length > 0) {
+    core.startGroup(`Removing temp folder ${tmpDir}`);
+    fs.rmdirSync(tmpDir, { recursive: true });
+    core.endGroup();
+  }
 }
